* origin/pr/110:
  qubes-vm-update: just print progress
  qubes-vm-update: plugins

+++ vmupdate/agent/source/plugins/manage_rpm_macro.py
@@ -20,24 +20,23 @@
 # USA.
 
 import os
-from typing import Dict, Optional
 
 import pkg_resources
 
 
-def manage_rpm_macro(os_data, log) -> Optional[Dict[str, str]]:
+def manage_rpm_macro(os_data, log, **kwargs):
     """
     Prepare requirements depend on os version.
     """
-    rpm_macro = "/usr/lib/rpm/macros.d/macros.qubes"
-    if (os_data["id"] == "fedora"
-            and os_data["release"] < pkg_resources.parse_version("33")):
-        log.info("Old fedora version detected.")
-        with open(rpm_macro, "w") as file:
-            file.write("# CVE-2021-20271 mitigation\n"
-                       "%_pkgverify_level all")
-    else:
-        if os.path.exists(rpm_macro):
-            os.remove(rpm_macro)
-        return {"dnf": "4.7.0", "rpm": "4.14.2"}
-    return None
+    if os_data["os_family"] == "RedHat":
+        rpm_macro = "/usr/lib/rpm/macros.d/macros.qubes"
+        if (os_data["id"] == "fedora"
+                and os_data["release"] < pkg_resources.parse_version("33")):
+            log.info("Old fedora version detected.")
+            with open(rpm_macro, "w") as file:
+                file.write("# CVE-2021-20271 mitigation\n"
+                           "%_pkgverify_level all")
+        else:
+            if os.path.exists(rpm_macro):
+                os.remove(rpm_macro)
+            kwargs["requirements"].update({"dnf": "4.7.0", "rpm": "4.14.2"})
